reordenar fix, eliminar ejercicios

+++ lib/screens/client/training/view_training_sessions/view_sesion_entrenamientos_screen.dart
@@ -67,7 +67,8 @@ class _ViewSesionEntrenamientoScreen
     await SesionEntrenamientoMethods().deleteSesionEntrenamiento(id);
   }
 
-  void _navigateNewSesion(SesionEntrenamiento sesionEntrenamiento) async {
+  void _navigateNewSesion(SesionEntrenamiento sesionEntrenamiento,
+      {bool isEdit = false}) async {
     Navigator.of(context)
         .push(MaterialPageRoute(
       builder: (context) => InfoSesionEntrenamientoScreen(
@@ -78,7 +79,8 @@ class _ViewSesionEntrenamientoScreen
     ))
         .then((deleteInfo) {
       initSesionEntrenamientos(); //cargo las sesiones y se actualizan
-      if (deleteInfo != null) {
+      if (deleteInfo != null && !isEdit) {
+        //debo borrar la sesión si es nueva y se ha quedado vacía
         _deleteSesionById(deleteInfo);
       }
     });
@@ -223,7 +225,7 @@ class _ViewSesionEntrenamientoScreen
   void _handleMenuItemSelected(String value, SesionEntrenamiento sesion) {
     switch (value) {
       case 'Editar':
-        _navigateNewSesion(sesion);
+        _navigateNewSesion(sesion, isEdit: true);
         break;
       case 'Eliminar':
         _deleteSesion(sesion);
